Fun looking PP page
I replaced everything with some images taken from the pp website
temporarily because it looks really cool.
It looks super awesome in portrait and in landscape it looks like they
are talking.
However, it is not appropriate for a preloader so it is going to be
reverted back in the next commit (which is soon).
I plan to upload it as a gist at a later time.