Fixed the ajax!!!
On a hunch, I finally figured out why the js code to change the user
thumbs doesn't work.
It turns out, it wasn't a problem with arrays, it wasn't a problem with
me and the DOM, and, it wasn't bad image urls or incorrect js, etc.
No, instead, it was a misplacement of three characters.
The problematic chars were "});".
Whew! I finally got it!