added a delay to the loading of the fullsize images
I realized that the javascript was changing the dom before the
thumbnails had a chance to load thus making them useless. To fix this I
added a small delay and reduced the thumbnail size; a delay of 900
milliseconds seems to be ideal delay considering both fast and slow
internet speeds.
When I am being less lazy, I will get the js to wait for the thumbnails
to finish loading then fetch the fullsize ones and swap them.