automatically load more gallery
Added automatic load next/prev page of gallery.
For the buttons, the transition is seamless, when using arrow keys, you continue to the overview like normal (but, the next/prev page of gallery has loaded).
Also, in the overview, the buttons should switch gallery pages without changing into a slide.