changed method of getting element from Hash
The new method uses the array of elements with class 'step' to figure out which element to return.
It expoits the fact that every slide has the 'step' class attached to it and that they are all in order in the DOM.
This method works with "6", or "step-6"; before only "step-6" would work.
This was actually coded a while ago but, held-back until now.