new mobile-focused version
About a week or two ago, I became frustrated with certain mobile browsers rendering of my gallery prototype.
I decided that it was time to make a lightweight version focused on mobile that was more or less written from scratch.
I never ended up commiting or pushing it b/c I was busy workinng on  various features.

I am finally pushing out a commit. This is a barebones image loader that supports the url scheme. It can probably be used well as a widget in many places if embedded.
However, b/c it was/is git-patch-staged from my current development version it may not actually work.