Urls for paintings not currently on the page will now load
Before, if a painting url was in the hash (url) that did not exist on the page, it would just get discarded.
Now, it will be fetched and navigated to.
That means that all painting urls can be entered manually now and that reloading with a painting url that's no longer on the first page now works instead of redirecting to the main gallery.