General improvements (tooltips, titles, cursor)
Page title now updates to reflect what is open and CSS powered tooltips have also been added so you know what you are pointing at.

When a painting is open, the page title will say: [painting name (or "artwork" if there isn't one)] by [name of painter].
Tooltips when hovering over a painting in the gallery have also been added using css that utilizes the data attribute; the nomenclature for these is the same as the title.
Data atributes are now being set for this info by javaScript.
Finally, the "pointer" cursor has been added to the css so it is more clear that the paintings are clickable.

Active slides will behave like they did before.