Show/hide user info with CSS instead of JS
The responsibility of showing and hiding user info has been transferred from the JS (DOM) to CSS based off of the gallery state.
This should improve the performance of the gallery since there is less code to run and also improves the consistency of the UI (the info sometimes didn't show before under certain urls due to buggy code).